backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.34.113
certifi==2025.4.26
click==8.2.0
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.115.12
google-generativeai==0.5.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
pymysql
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
sqlalchemy>=2.0
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
cryptography
